Move log writing to a separate thread [VPNAND-451].
ProtonLoggerImpl uses a logging library that writes to log file
synchronously. The logging code is moved to a separate, dedicated thread.
This achieves two goals:
- adding logs on the main thread doesn't block it for disk I/O,
- logger is single-threaded which should make it easier to implement
  functionality like clearing logs.

Additionally getLogLines() loads log data on the logger thread which
avoids blocking the main thread when logs are about to be displayed in
LogFragment.

The communication between logging threads and the logger thread is done
via a MutableSharedFlow with a limited buffer. When the buffer gets full
the most recent messages are dropped.