Copy log files for upload [VPNAND-451].
When a bug report is being uploaded with log reports, the log files are
being read by OkHttp's worker thread during data transfer. Therefore it
is possible for logs to be appended during this process which results in
errors because the transferred content length is different than that
declared in HTTP header.
With this commit log files are copied to temporary files for upload
ensuring that their content doesn't change due to messages being
appended of files being rotated during upload.