Add ProfileColor enum and update ProfileActivity [VPNAND-505].
Introduce ProfileColor to represent the palette of profile colors. The
previously used color strings are migrated to the palette on load.

The custom ProtonPalette view is removed, instead the grid of color
selectors is created dynamically based on the ProfileColor enum.

The ProfileActivity is slightly refactored with some functionality moved
to ProfileViewModel.