Put all texts in layout to ensure consistency [VPNAND-540].
All pager fragments need to be of the same height, otherwise they may be
cut at the bottom (unless the first fragment is the largest).
The texts can vary in length between languages (and also the cog wheel
icon takes some additional vertical space) so the simplest solution is
to render all of them in each fragment to reserve space.
Only one is ever visible.