Fix excluded IPs header updates [VPNAND-542].
Updating header via Section.setHeader() triggers item notifications for
remove and a subsequent add which animates the item to first disappear
and then appear.
We need an "item changed" notification so that the item is updated
instantly (animation of change updates is disabled).
The fix is to update the whole section in the GroupAdapter and rely on
the HeaderViewHolder's ID being fixed.