Don't block settings when connected [VPNAND-597].
Settings that affect VPN connections are no longer blocked while
connected.

For switches the user is asked if they want to reconnect.

For complex settings with their own screen the data is saved and after
that the user is shown a dialog with information that the changes will
be applied on next connection. The user can reconnect from the dialog.