Migrate test code to Hilt [VPNAND-614]
This is initial change to get our tests working with Hilt but it needs to be
followed-up by more hilt-friendly structure.

Main problem is that our old code was injecting objects to fields (including
static fields) which lifecycle spanned multiple test case which is against
hilt's approach of testing where SingletonComponent have a lifetime of a single
test case. This means that:
- @Rule (that lives though multiple test cases) that's using injected fields
needs to re-initialize them for each test case.
- @ClassRule can't use injected fields at all
- Robots that use injected fields needs to be re-initialized for each test case