Fix bottom sheet not opening sometimes in tests.
Animating the peek height at the same time as animating expand can
abort the expand (possibly a bug in BottomSheetBehavior). It's possible
this only occurs in automated tests when a mocked connection changes
several states immediately.
The animation is disabled to fix the issue. The bottom sheet's and peek
height should change only slightly so it doesn't actually need
animating.