Separate methods for connecting in fg and in bg.
This way VpnConnectionManager makes fewer decisions based on the
type of Context it is provided. This was error prone and became even
more problematic with the introduction of Hilt that uses wrappers on
certain contexts (e.g. Fragment.requireContext() no longer returns an
Activity).
The new connect method accepts a VpnPermissionDelegate to avoid using an
Activity directly (an Activity is problematic to mock in some tests).