Fix report form being shown after submission [VPNAND-690].
The loading UI was being managed by the code making the request which
would result in the UI being closed independently of the
ReportButActivity finishing itself. This was leading to the form being
visible for a brief moment before the activity was finished.

With this change it is the activity that manages the loading UI
(in response to updates from the ViewModel). When the report is submitted
successfully the activity is finished without removing the loading UI so
the form doesn't reappear.