Update server status on the unfiltered list [VPNAND-404].
allServers is changed to reference servers from the unfiltered lists of
countries.
The filtered collections are being rebuilt from the unfiltered ones so
any modifications done to them are lost on an update.

allServers is used to update status of individual servers, like load and
score, and therefore it needs to be based on the unfiltered collections.