Use up-to-date server object to get load [VPNAND-404].
Reloading the full server list creates new Server objects to represent
the servers. Any referenced Servers, e.g. in ConnectionParams, become
stale.
This isn't super bad because most properties of a server never change.
To get the current load or score, an up-to-date Server needs to be
retrieved from the ServerManager.