Open VPN settings in a new task.
VPN settings allow navigating to our VPN app by clicking the app entry.
However, when the settings screen is part of tha application's task,
navigating to it opens the exact same screen that is already displayed.
This is confusing.
The VPN settings activity is now being opened in a new task.