Decouple ProtonLogger from logging implementation.
ProtonLogger is used all over the code base for easy logging. This
includes code that is being tested with unit tests. A useful logging
implementation needs access to application's Context to be able to write
files in app folders etc. An application Context is not available in
unit tests.
With this change ProtonLogger delegates to a logging implementation, a
fake one in unit tests and a real one when the whole application is
started (including UI tests).