Fix excluded apps viewState flow when resubscribed [VPNAND-791].
When the UI is stopped and started again it starts collecting the
viewState flow again. This flow combines a number of other flows that
share their state in the viewModelScope. Some of these flows did not
reemit the last value resulting in the combine operator to wait for
input indefinitely.
The fix is to set all the intermediate flows to replay their last value.