Use WorkManager to scheduler cert refresh [VPNAND-796].
ReschedulableTask used for certificate refresh uses Handler.postDelayed
under the hood. This means that the set delay does not account for the
time the device is in deep sleep, making it unsuitable for scheduling
tasks to be run after many hours.

With this commit WorkManager is used to schedule the refresh task.

AppInUseMonitor is introduced to avoid scheduling refresh when the app
is no longer in active use,