Introduce a way to inject "self-contained" objects.
Some classes, e.g. the MaintenanceTracker or AccountStateHandler, are
not used by any other parts of the app - they observe some state and
react to its changes.
They are not automatically instantiated by Hilt because there is nowhere
to inject them to.
Some of them are injected to ProtonApplicationHilt but this means they
are not instantiated in tests.

This change introduces ProtonApplication.initDependencies() method and
ProtonHiltApplicationRule for initializing such objects also in tests.