SwitchDialogActivity opens activities in same task [VPNAND-811].
The action for the main SwitchDialogActivity's button was defined as a
PendingIntent. This means that if it opened an Activity, the Activity
was being be placed in a new task (because taskAffinity="").

To prevent this the ActionItem is extended to either provide an Intent
for an Activity or a PendingIntent for other actions.
This allows SwitchDialogActivity to open the next Activity in the same
task.

Activities opened via notification will be put in a separate task but
this is acceptable because they represent a flow separate of the main
UI.