Setting NetShieldSwitch programmatically doesn't update settings.
NetShieldSwitch calls a callback whenever the UI switch is toggled. The
callback updates the NetShield protocol in UserData. However, the value
in UserData can be null, while the UI swich is only true or false.
Setting the UserData value to null may propagate to the UI which in turn
updates the UserData value with a non-null setting.
This is worked around by disabling the switch's change listener while
updating its position programmatically.