Compare by server ID when saving to profiles.
Save to profile checks if there already is a profile for the current
server by comparing server's domain. However, several servers can share
a single domain which makes it impossible to save more than one server
with the domain. This is confusing to users - to them these are distinct
servers.
Use serverID for comparison instead.