Reconnect on app startup [VPNAND-869].
If the app is stopped (e.g. due to memory pressure) while VPN is active
the VPN service should be restarted by the OS because our VPN services
return START_STICKY.
Apparently this doesn't always happen.
Restore VPN connection on app start to handle the edge cases where the
process gets restarted but not the service.

The stored ConnectionParams is being cleared when VPN state reaches
Disabled, not in the disconnect* methods. ConnectionParams is extended
with a UUID to make sure the proper object is deleted.