Prevent device sleep when recovering connection [VPNAND-888].
Recovering from a VPN connection error may require making several API
calls, pinging some servers etc. When it happens while the app is in
background there's a possibility of the device entering sleep mode in
the middle of the process.
To prevent this a wake lock is kept during error handling and server
ping.