Make CurrentUser test friendly.
The CurrentUser class is mocked in many tests to provide a user. Mocking
requires knowing which methods to mock and how which makes writing tests
difficult.

With this change a CurrentUserProvider is factored out of the
CurrentUser class. This allows using the real CurrentUser implementation
in tests with a simple TestCurrentUserProvider and no mocking
whatsoever.