Ensure server list is loaded for certain screens [VPNAND-1517].
Some screens assume the server list is always in memory instead of
observing it for changes (esp. screens that show details for a specific
country). They work most of the time but can fail or crash when restored
after process death.

They are updated to wait for the list to be loaded.