Pass params UUID to onVpnServiceDestroyed.
onVpnServiceDestroyed deletes ConnectionParams for the active
connection. This can lead to deleting the wrong params when reconnecting
with some unfortunate timing.
To mitigate this OpenVPNWrapperService tracks the last known
ConnectionParams UUID and passes it to onVpnServiceDestroyed.

Not implemented for Wireguard because we don't control the intent for
starting the service.