Don't auto-toggle Custom DNS settings in settings manager [VPNAND-2125].
CurrentUserLocalSettingsManager used to change the enabled status of
custom DNS depending on the DNS list updates.
This logic has been removed, the toggle is only changed by the settings
view model when adding a new DNS item.
The `enabled` field is renamed to `switchEnabled` to explicitly state
it's about the UI toggle, not feature as a whole.
`effectiveEnabled` should be used in most places..