Fix SC arc cut off for greyed out servers [VPNAND-2223].
The arc is being drawn outside the layout bounding box of the flag
composable. This simplifies positioning a lot and works well in most
cases.
Applying alpha moves the composable to a separate graphics layer and
it's impossible to draw beyond its bounds.
The fix is to apply alpha to the whole row instead of its individual
components separately. This way the whole row is in its own layer,
together with padding, which leaves space for drawing the arc.